requirements.txt
requests
irc
aiohttp
dotmap
PyYAML
